backend/requirements.txt
bcrypt
fastapi
SQLAlchemy
pytest
uvicorn
httpx
python-dateutil
